chore: move to dto modelss
