+++ api/app/src/main/kotlin/packit/model/dto/PacketGroupSummary.kt
@@ -1,0 +1,10 @@
+package packit.model.dto
+
+// Projection class for PacketRepository.findPacketGroupSummaryByName
+interface PacketGroupSummary
+{
+    fun getName(): String
+    fun getPacketCount(): Int
+    fun getLatestId(): String
+    fun getLatestTime(): Double
+}
